Embed mruby to master

+++ mruby_components/mruby/mrbgems/mruby-exit/src/mruby-exit.c
@@ -1,5 +1,5 @@
 #include <stdlib.h>
-#include "mruby.h"
+#include <mruby.h>
 
 static mrb_value
 f_exit(mrb_state *mrb, mrb_value self)
@@ -7,7 +7,7 @@ f_exit(mrb_state *mrb, mrb_value self)
   mrb_int i = EXIT_SUCCESS;
 
   mrb_get_args(mrb, "|i", &i);
-  exit(i);
+  exit((int)i);
   /* not reached */
   return mrb_nil_value();
 }
